OlgaVarivonchik/requirements.txt
argparse
bs4
datetime
requests
python-dateutil
fpdf
lxml
colorama